Optimize getting caller location using some ruby internals
We want to get the caller's control frame path and line number in
a trace event hook without the overhead of going through a method
like `caller` or `caller_locations`.  The ruby tracepoint API only
gives access to the called functions path and line number.

Instead, we rely on the minimum amount or ruby internals needed to
get the caller's path and line number by changing the `cfp` field
of the internal rb_trace_arg_struct before getting the path and
lineno from the trace arg.