Fix block method return filepath and lineno using call stack frame.
The problems with the filepath and lineno for block methods only
happens on the return event, so use the filepath and lineno from
the call event on the return event to avoid this problem.  Since we
use the filepath for filtering, we also need to include blacklisted
call frames that aren't logged on the rotoscope stack.