Use direct caller in flattened output even if its callsite is blacklisted (#54)
If the caller frame is blacklisted, then that means its caller was from a
blacklisted file path, which isn't relevant for the current call. There
is no reason to iterate through the callers to find a non-blacklisted one
since we already have an early return if the call site for the current
call is in the blacklist.