Reorder CSV fields to simplify use through an awk script
The method name might have commas in it, so using a comma as a field
separator in an awk script would not split it properly. We care about
the entity and filepath fields in that awk script but not the method
name, so I've moved the method names to come afterwards.